refactor: extract RecordingStatusMessageSnapshot.swift from RecordingStatusMessageProvider.swift
Splits the `RecordingStatusMessageSnapshot` value struct out of
RecordingStatusMessageProvider.swift so that file holds only the
provider class + the PostTranscriptionStatusPresenter class.

Byte-identical move of lines 4-10. New file needs only Foundation
(the provider file keeps Combine + Foundation).

RecordingStatusMessageProvider.swift: 100 → 92 lines. Tests: 667.
