+++ Sources/BugNarrator/Utilities/RecordingStatusMessageProvider.swift
@@ -1,14 +1,6 @@
 import Combine
 import Foundation
 
-struct RecordingStatusMessageSnapshot: Equatable {
-    var audioSource: RecordingAudioSource
-    var hasUsableAIProviderCredential: Bool
-    var aiProviderCompatibilityIssue: String?
-    var autoExtractIssues: Bool
-    var autoCopyTranscript: Bool
-}
-
 final class RecordingStatusMessageProvider {
     private let snapshot: () -> RecordingStatusMessageSnapshot
